Break apart editor context
