+++ plugins/rich-editor/src/scripts/editor/pieces/EditorUploadButton.tsx
@@ -8,7 +8,8 @@ import React from "react";
 import classNames from "classnames";
 import { attachment, emoji, image } from "@library/icons/editorIcons";
 import EmbedInsertionModule from "@rich-editor/quill/EmbedInsertionModule";
-import { IWithEditorProps, withEditor } from "@rich-editor/editor/context";
+import { IWithEditorProps } from "@rich-editor/editor/context";
+import { withEditor } from "@rich-editor/editor/withEditor";
 import { isFileImage } from "@library/utility/utils";
 import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
 import { IconForButtonWrap } from "@rich-editor/editor/pieces/IconForButtonWrap";
