+++ plugins/rich-editor/src/scripts/flyouts/EmbedFlyout.tsx
@@ -14,7 +14,8 @@ import FrameBody from "@library/layout/frame/FrameBody";
 import FrameFooter from "@library/layout/frame/FrameFooter";
 import { isAllowedUrl, t } from "@library/utility/appUtils";
 import { getRequiredID, IRequiredComponentID } from "@library/utility/idUtils";
-import { IWithEditorProps, withEditor } from "@rich-editor/editor/context";
+import { IWithEditorProps } from "@rich-editor/editor/context";
+import { withEditor } from "@rich-editor/editor/withEditor";
 import { IconForButtonWrap } from "@rich-editor/editor/pieces/IconForButtonWrap";
 import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
 import { insertMediaClasses } from "@rich-editor/flyouts/pieces/insertMediaClasses";
