+++ plugins/rich-editor/src/scripts/flyouts/pieces/EmojiPicker.tsx
@@ -9,7 +9,8 @@ import { AutoSizer, Grid } from "react-virtualized";
 import classNames from "classnames";
 import { t } from "@library/utility/appUtils";
 import { IFlyoutToggleChildParameters } from "@library/flyouts/FlyoutToggle";
-import { IWithEditorProps, withEditor } from "@rich-editor/editor/context";
+import { IWithEditorProps } from "@rich-editor/editor/context";
+import { withEditor } from "@rich-editor/editor/withEditor";
 import { EMOJIS, EMOJI_GROUPS } from "@rich-editor/flyouts/pieces/emojiData";
 import { emojiGroupsClasses } from "@rich-editor/flyouts/pieces/insertEmojiGroupClasses";
 import { insertEmojiClasses } from "@rich-editor/flyouts/pieces/insertEmojiClasses";
