+++ plugins/rich-editor/src/scripts/flyouts/pieces/Flyout.tsx
@@ -8,7 +8,8 @@ import React from "react";
 import classNames from "classnames";
 import CloseButton from "@library/navigation/CloseButton";
 import { flyoutPosition } from "@rich-editor/flyouts/pieces/flyoutPosition";
-import { IWithEditorProps, withEditor } from "@rich-editor/editor/context";
+import { IWithEditorProps } from "@rich-editor/editor/context";
+import { withEditor } from "@rich-editor/editor/withEditor";
 import { richEditorFlyoutClasses } from "@rich-editor/flyouts/pieces/flyoutClasses";
 import { richEditorClasses } from "@rich-editor/editor/richEditorClasses";
 import { ButtonTypes } from "@library/forms/buttonStyles";
