+++ plugins/rich-editor/src/scripts/toolbars/MentionToolbar.tsx
@@ -9,7 +9,8 @@ import Quill, { Sources, DeltaStatic } from "quill/core";
 import uniqueId from "lodash/uniqueId";
 import isEqual from "lodash/isEqual";
 import Keyboard from "quill/modules/keyboard";
-import { withEditor, IWithEditorProps } from "@rich-editor/editor/context";
+import { IWithEditorProps } from "@rich-editor/editor/context";
+import { withEditor } from "@rich-editor/editor/withEditor";
 import MentionAutoCompleteBlot from "@rich-editor/quill/blots/embeds/MentionAutoCompleteBlot";
 import { getBlotAtIndex } from "@rich-editor/quill/utility";
 import { connect } from "react-redux";
